WIP: Refactor simulation architecture. Key changes include:
- Separate SimulationEnvironment class for core physics simulation
- Add TimeStepController for consistent timing and real-time control
- Implement pub/sub pattern for simulation event notifications
- Move robot functionality to dedicated Robot class
